+++ lib/helper.py
@@ -1,0 +1,7 @@
+import socket
+import struct
+def cidr_to_netmask(cidr):
+    network, net_bits = cidr.split('/')
+    host_bits = 32 - int(net_bits)
+    netmask = socket.inet_ntoa(struct.pack('!I', (1 << 32) - (1 << host_bits)))
+    return network, netmask
